requirements.txt
engineering-notation==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
